Automatically truncate sequences for decoder-only
Expecting the user to supply a sequence length greater than any data
point is ridiculous. So now we greedily truncate the sequence based on
the maximum amount of `extra_id`s, which wastes a lot of data. An
alternative would be going a statistical route with significance
attached to it; allowing the expected amount of tokens with some leeway,
while handling an unlikely length excession error.

This only handles the decoder-only case, while the encoder-decoder case
is left as is. This is because errors are much less like for the
encoder-decoder case unless massive corruption is configured or if the
decoder has a smaller sequence length than the encoder.